refactor(test): move test helpers into testHelpers dir
